Move serve::Listener into its own module
